Attempt at synchronizing web medals with in-game

+++ main.go
@@ -419,11 +419,11 @@ where hash = $1`, hash).Scan(&de, &de2, &dap, &daw, &dal, &dui, &dallowed, &dep,
 			if dal == 0 {
 				dal = 1
 			}
-			if daw >= 24 && daw/dal > 12 {
+			if daw >= 24 && float64(daw)/float64(dal) > 12.0 {
 				m.Medal = 1
-			} else if daw >= 12 && daw/dal > 6 {
+			} else if daw >= 12 && float64(daw)/float64(dal) > 6.0 {
 				m.Medal = 2
-			} else if daw >= 6 && daw/dal > 3 {
+			} else if daw >= 6 && float64(daw)/float64(dal) > 3.0 {
 				m.Medal = 3
 			}
 			if de > 1800 {
